fix add-to-list component
