+++ src/components/add-to-list/add-to-list.jsx
@@ -1,9 +1,6 @@
 import React from "react";
-import {connect} from "react-redux";
 import PropTypes from "prop-types";
 import {AuthorizationStatus, AppRoute} from "../../const.js";
-import {Operation as DataOperation} from '../../reducer/data/data.js';
-import {getAuthorizationStatus} from "../../reducer/user/selectors.js";
 import history from "../../history.js";
 
 const redirectToSignIn = (authorizationStatus, onMyListBtnClick, id, isFavorite) => {
@@ -51,15 +48,4 @@ AddToList.propTypes = {
   authorizationStatus: PropTypes.string.isRequired,
 };
 
-const mapStateToProps = (state) => ({
-  authorizationStatus: getAuthorizationStatus(state),
-});
-
-const mapDispatchToProps = (dispatch) => ({
-  onMyListBtnClick(id, isFavorite) {
-    dispatch(DataOperation.setFavorite(id, isFavorite));
-  }
-});
-
-export {AddToList};
-export default connect(mapStateToProps, mapDispatchToProps)(AddToList);
+export default AddToList;
