+++ src/components/promo-section/promo-section.jsx
@@ -6,7 +6,7 @@ import history from "../../history.js";
 import ErrorMessage from "../error-message/error-message.jsx";
 
 import UserBlock from "../user-block/user-block.jsx";
-import AddToList from "../add-to-list/add-to-list.jsx";
+import AddToList from "../add-to-list/add-to-list.connect.js";
 
 const PromoSection = (props) => {
   const {film, loadPromoError} = props;
